Convert Problem5 to use longs instead of ints.
Ints aren't large enough to handle 20 factorial.
This added a slight complication in that Enumerable.Range is not available
to use with longs, so I had to roll my own.

+++ Tests/Problem5Tests.cs
@@ -16,7 +16,7 @@ namespace Tests {
         }
 
         [TestCaseSource("GetSomeNumbersAndTheirExpectedFactorials")]
-        public void ShouldReturnTheCorrectFactorials (int number, int expectedFactorial) {
+        public void ShouldReturnTheCorrectFactorials (long number, long expectedFactorial) {
             Assert.AreEqual(expectedFactorial, number.Factorial());
         }
